+++ backend/static/_app/immutable/chunks/BsCIOrMO.js
@@ -1,1 +1,0 @@
-import{M as o,b as u,ag as g,o as t,Q as _,Y as p,a9 as i,W as l,O as d,X as m}from"./-s7GvDYA.js";function E(n,r){let a=null,f=t;var s;if(t){a=m;for(var e=_(document.head);e!==null&&(e.nodeType!==p||e.data!==n);)e=i(e);if(e===null)l(!1);else{var y=i(e);e.remove(),d(y)}}t||(s=document.head.appendChild(o()));try{u(()=>r(s),g)}finally{f&&(l(!0),d(a))}}export{E as h};